frontend empty chart data, warning on edit/add endpoint with no resp codes

+++ frontend/src/pages/collector/endpoints/Dialog.tsx
@@ -24,41 +24,43 @@ export function Dialog(props: DialogProps) {
     return (
         props.isOpen && <AlertDialog>
             <AlertDialog.Backdrop isOpen={props.isOpen} onOpenChange={props.setIsOpen}>
-                <AlertDialog.Dialog>
-                    <AlertDialog.Header>
-                        <h4>{title} Endpoint{props.action === "delete" && "?"}</h4>
-                        <AlertDialog.CloseTrigger/>
-                    </AlertDialog.Header>
-                    {
-                        props.action !== "delete"
-                            ? <AlertDialog.Body>
-                                <EndpointsForm
-                                    action={props.action}
-                                    endpoint={ props.action === "edit" ? props.endpoint : undefined}
-                                    collectorId={props.collector_id}
-                                    setIsOpen={props.setIsOpen}
-                                    refresh={props.refresh}
-                                />
-                            </AlertDialog.Body>
-                            : <>
-                                <AlertDialog.Body>
-                                    <p>Collector will no longer send requests to this endpoint</p>
-                                    <p>This action is not reversible</p>
+                <AlertDialog.Container size={"lg"}>
+                    <AlertDialog.Dialog>
+                        <AlertDialog.Header>
+                            <h4>{title} Endpoint{props.action === "delete" && "?"}</h4>
+                            <AlertDialog.CloseTrigger/>
+                        </AlertDialog.Header>
+                        {
+                            props.action !== "delete"
+                                ? <AlertDialog.Body>
+                                    <EndpointsForm
+                                        action={props.action}
+                                        endpoint={ props.action === "edit" ? props.endpoint : undefined}
+                                        collectorId={props.collector_id}
+                                        setIsOpen={props.setIsOpen}
+                                        refresh={props.refresh}
+                                    />
                                 </AlertDialog.Body>
-                                <AlertDialog.Footer>
-                                    <Button slot={"close"} variant={"tertiary"}>Cancel</Button>
-                                    <Button
-                                        slot={"close"}
-                                        variant={"danger"}
-                                        onClick={() => {
-                                            deleteEndpoint(props.endpoint?.id || 0)
-                                            props.refresh()
-                                        }}
-                                    >Delete</Button>
-                                </AlertDialog.Footer>
-                            </>
-                    }
-                </AlertDialog.Dialog>
+                                : <>
+                                    <AlertDialog.Body>
+                                        <p>Collector will no longer send requests to this endpoint</p>
+                                        <p>This action is not reversible</p>
+                                    </AlertDialog.Body>
+                                    <AlertDialog.Footer>
+                                        <Button slot={"close"} variant={"tertiary"}>Cancel</Button>
+                                        <Button
+                                            slot={"close"}
+                                            variant={"danger"}
+                                            onClick={() => {
+                                                deleteEndpoint(props.endpoint?.id || 0)
+                                                props.refresh()
+                                            }}
+                                        >Delete</Button>
+                                    </AlertDialog.Footer>
+                                </>
+                        }
+                    </AlertDialog.Dialog>
+                </AlertDialog.Container>
             </AlertDialog.Backdrop>
         </AlertDialog>
     )
